integration tests: Move smbclient pod creation to util_test
